feat(FINAL): Make admin cannot delete all banners
- Modify on client:
  - src/pages/Banner/:
    - index.jsx: Add prop `isOneLeft` to BannerRow
    - components/BannerRow/index.jsx: Return toast.error when admin try
      to delete the remaining one
    - src/languages: Add one translation sentence
  - __tests__/pages/Banner/components/BannerRow/index.test.jsx: Add
    props `isOneLeft`

- Modify on server:
  - bannerController.js: Return 400 when admin try to delete the remaining one
  - handleJob.test.js: Increase the timeout value

+++ server/controllers/bannerController.js
@@ -86,6 +86,9 @@ exports.deleteBanner = async (req, res) => {
     const foundBanner = await Banner.findByPk(bannerId);
     if (!foundBanner) return handleClientError(res, 404, 'Banner Not Found');
 
+    const banners = await Banner.findAll();
+    if (banners.length === 1) return handleClientError(res, 400, 'Should have minimum 1 banner');
+
     await Banner.destroy({ where: { id: bannerId } });
     fs.unlinkSync(foundBanner.imageDesktop);
     fs.unlinkSync(foundBanner.imageMobile);
